# backend/gen/api/students/schema.gen.go
// Package students_gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package students_gen

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreateStudentRequest defines model for CreateStudentRequest.
type CreateStudentRequest struct {
	Name       string `json:"name"`
	RollNumber string `json:"roll_number"`
}

// Student defines model for Student.
type Student struct {
	ClassId    openapi_types.UUID `json:"class_id"`
	Id         openapi_types.UUID `json:"id"`
	Name       string             `json:"name"`
	RollNumber string             `json:"roll_number"`
}

// StudentList defines model for StudentList.
type StudentList struct {
	Students []Student `json:"students"`
}

// UpdateStudentRequest defines model for UpdateStudentRequest.
type UpdateStudentRequest struct {
	Name       string `json:"name"`
	RollNumber string `json:"roll_number"`
}

// CreateStudentJSONRequestBody defines body for CreateStudent for application/json ContentType.
type CreateStudentJSONRequestBody = CreateStudentRequest

// UpdateStudentJSONRequestBody defines body for UpdateStudent for application/json ContentType.
type UpdateStudentJSONRequestBody = UpdateStudentRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List Students in a Class
	// (GET /classes/{classId}/students)
	ListStudents(c *gin.Context, classId openapi_types.UUID)
	// Add a Student to a Class
	// (POST /classes/{classId}/students)
	CreateStudent(c *gin.Context, classId openapi_types.UUID)
	// Remove a Student from a Class
	// (DELETE /classes/{classId}/students/{studentId})
	DeleteStudent(c *gin.Context, classId openapi_types.UUID, studentId openapi_types.UUID)
	// Get a Student in a Class
	// (GET /classes/{classId}/students/{studentId})
	GetStudent(c *gin.Context, classId openapi_types.UUID, studentId openapi_types.UUID)
	// Update a Student in a Class
	// (PUT /classes/{classId}/students/{studentId})
	UpdateStudent(c *gin.Context, classId openapi_types.UUID, studentId openapi_types.UUID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// ListStudents operation middleware
func (siw *ServerInterfaceWrapper) ListStudents(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListStudents(c, classId)
}

// CreateStudent operation middleware
func (siw *ServerInterfaceWrapper) CreateStudent(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateStudent(c, classId)
}

// DeleteStudent operation middleware
func (siw *ServerInterfaceWrapper) DeleteStudent(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "studentId" -------------
	var studentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "studentId", c.Param("studentId"), &studentId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter studentId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteStudent(c, classId, studentId)
}

// GetStudent operation middleware
func (siw *ServerInterfaceWrapper) GetStudent(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "studentId" -------------
	var studentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "studentId", c.Param("studentId"), &studentId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter studentId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetStudent(c, classId, studentId)
}

// UpdateStudent operation middleware
func (siw *ServerInterfaceWrapper) UpdateStudent(c *gin.Context) {

	var err error

	// ------------- Path parameter "classId" -------------
	var classId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "classId", c.Param("classId"), &classId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter classId: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Path parameter "studentId" -------------
	var studentId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "studentId", c.Param("studentId"), &studentId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter studentId: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateStudent(c, classId, studentId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/classes/:classId/students", wrapper.ListStudents)
	router.POST(options.BaseURL+"/classes/:classId/students", wrapper.CreateStudent)
	router.DELETE(options.BaseURL+"/classes/:classId/students/:studentId", wrapper.DeleteStudent)
	router.GET(options.BaseURL+"/classes/:classId/students/:studentId", wrapper.GetStudent)
	router.PUT(options.BaseURL+"/classes/:classId/students/:studentId", wrapper.UpdateStudent)
}